+++ lib/src/iterable_extensions.dart
@@ -1,4 +1,5 @@
-import 'collection_util.dart';
+import 'package:collection/collection.dart' as coll_lib;
+
 import 'collections.dart' as coll;
 import 'objects.dart';
 
@@ -10,6 +11,7 @@ extension IterableExtensions<E> on Iterable<E> {
   /// empty
   ///
   /// `since 0.0.1`
+  //TODO move to nullable extension
   E? get firstOrNull {
     return coll.firstOrNull(this);
   }
@@ -33,14 +35,6 @@ extension IterableExtensions<E> on Iterable<E> {
     return condition ? where(predicate) : this;
   }
 
-  /// Takes a [path] `List` and returns the value at [path] in a nested collection.
-  /// Traverses `Map` (by key), `List` (by index), and `Iterable` (by index) values.
-  ///
-  /// `since 0.5.0`
-  dynamic getIn(List<dynamic> path) {
-    return CollectionUtil.getIn(path, this);
-  }
-
   /// Returns an [Iterable] of [Iterable]s of [n] items each without overlap.
   ///
   /// `since 0.6.0`
@@ -48,6 +42,15 @@ extension IterableExtensions<E> on Iterable<E> {
     assert(0 < n);
     // TODO Consider making lazy
     return isEmpty ? [] : ([take(n), ...skip(n).partition(n)]);
+  }
+
+  /// Groups the elements by the value returned by [key].
+  ///
+  /// Returns a map from keys computed by [key] to a list of all values for which
+  /// [key] returns that key. The values appear in the list in the same relative
+  /// order as [this].
+  Map<K, List<E>> groupBy<K>(K Function(E) key) {
+    return coll_lib.groupBy(this, key);
   }
 }
 
@@ -59,7 +62,21 @@ extension NestedIterableExtensions<E> on Iterable<Iterable<E>> {
   /// elements in iteration order.
   ///
   /// `since 0.0.1`
-  Iterable<E?> get flatten {
+  Iterable<E> get flatten {
     return expand(identity);
   }
+}
+
+/// Extends [Iterable] of [Future]s
+///
+/// `since 0.8.0`
+extension FutureIterableExtensions<V> on Iterable<Future<V>> {
+  /// Does a [Future.wait] for all values in [this].
+  ///
+  /// `since 0.8.0`
+  Future<List<V>> awaitAll({
+    bool eagerError = false,
+    void Function(V successValue)? cleanUp,
+  }) =>
+      Future.wait(this, eagerError: eagerError, cleanUp: cleanUp);
 }