+++ lib/src/map_extensions.dart
@@ -1,4 +1,3 @@
-import 'collection_util.dart';
 import 'map_util.dart';
 
 /// Useful Map extensions
@@ -10,13 +9,5 @@ extension MapExtentions on Map {
   /// `since 0.4.0`
   Map removeNullValues() {
     return MapUtil.removeNullValues(this);
-  }
-
-  /// Takes a [path] `List` and returns the value at [path] in a nested collection.
-  /// Traverses `Map` (by key), `List` (by index), and `Iterable` (by index) values.
-  ///
-  /// `since 0.5.0`
-  dynamic getIn(List<dynamic> path) {
-    return CollectionUtil.getIn(path, this);
   }
 }