+++ lib/src/string_extensions.dart
@@ -5,14 +5,14 @@ import 'strings.dart' as StringUtils;
 ///
 /// `since 0.7.0`
 extension NullableStringExtensions on String? {
-  /// Returns `true` if `this` is NOT blank. See [StringUtils.isBlank]
+  /// Returns `true` if [this] is NOT blank. See [StringUtils.isBlank]
   ///
   /// `since 0.7.0`
   bool get isNotBlank {
     return StringUtils.isNotBlank(this);
   }
 
-  /// Returns `true` if `this` is blank. See [StringUtils.isBlank]
+  /// Returns `true` if [this] is blank. See [StringUtils.isBlank]
   ///
   /// `since 0.7.0`
   bool get isBlank {