+++ test/iterable_extensions_test.dart
@@ -21,7 +21,7 @@ void main() {
     var collection1 = ['a', 'b'];
     var collection2 = [1, 2];
 
-    Iterable<Object?> actual = [collection1].flatten;
+    Iterable<Object> actual = [collection1].flatten;
     expect(actual, equals(collection1));
 
     actual = [collection1, collection2].flatten;
@@ -99,5 +99,21 @@ void main() {
       var actual = coll.filterWhen(true, (element) => element.isEven);
       expect(actual, equals([2, 4]));
     });
+  });
+
+  test('groupBy', () {
+    var a = [1, 'a'];
+    var b = [2, 'b'];
+    var c = [1, 'c'];
+    var coll = [a, b, c];
+    expect(coll.groupBy((v) => v.first), {
+      1: [a, c],
+      2: [b]
+    });
+  });
+
+  test('awaitAll', () async {
+    var coll = [1, 2, 3];
+    expect(await coll.map(Future.value).awaitAll(), coll);
   });
 }